Move SSH terminal tests into terminal suite
