+++ Tests/QuillCodeAppTests/WorkspaceTerminalIntegrationTests.swift
@@ -1,5 +1,7 @@
 import Foundation
 import XCTest
+import QuillCodeCore
+import QuillCodeTools
 @testable import QuillCodeApp
 
 @MainActor
@@ -26,6 +28,85 @@ final class WorkspaceTerminalIntegrationTests: XCTestCase {
         XCTAssertEqual(surface.cwdLabel, root.path)
         XCTAssertEqual(surface.entries.first?.statusLabel, "Done")
         XCTAssertEqual(surface.entries.first?.exitCodeLabel, "exit 0")
+    }
+
+    func testTerminalCommandRunsThroughSSHRemoteProject() async throws {
+        let root = try makeTempDirectory()
+        let argumentsFile = root.appendingPathComponent("ssh-args.txt")
+        let fakeSSH = try makeFakeSSH(in: root, argumentsFile: argumentsFile)
+        let connection = ProjectConnection.ssh(
+            path: "/srv/quill repo",
+            host: "feather.local",
+            user: "quill",
+            port: 2222
+        )
+        let project = ProjectRef(name: "Feather", path: connection.path, connection: connection)
+        let model = QuillCodeWorkspaceModel(
+            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
+            sshRemoteShellExecutor: SSHRemoteShellExecutor(
+                sshExecutable: fakeSSH.path,
+                connectTimeoutSeconds: 4
+            )
+        )
+
+        await model.runTerminalCommand("printf remote-terminal", workspaceRoot: root)
+
+        XCTAssertEqual(model.terminal.entries.count, 1)
+        XCTAssertEqual(model.terminal.entries[0].status, .done)
+        XCTAssertEqual(model.terminal.entries[0].stdout, "remote-terminal\n")
+        XCTAssertEqual(model.terminal.currentDirectoryPath, "ssh://quill@feather.local:2222/srv/quill repo")
+        let surface = model.surface()
+        XCTAssertEqual(surface.terminal.cwdLabel, "ssh://quill@feather.local:2222/srv/quill repo")
+        XCTAssertEqual(surface.terminal.entries.first?.executionContext?.kind, .sshRemote)
+        XCTAssertEqual(surface.terminal.entries.first?.executionContext?.label, "SSH Remote")
+        XCTAssertEqual(surface.terminal.entries.first?.executionContext?.detail, "feather.local")
+        let arguments = try String(contentsOf: argumentsFile, encoding: .utf8)
+        XCTAssertTrue(arguments.contains("-T\n-o\nBatchMode=yes\n-o\nConnectTimeout=4\n-p\n2222\nquill@feather.local\n"))
+        XCTAssertTrue(arguments.contains("cd '/srv/quill repo' &&"))
+        XCTAssertTrue(arguments.contains("printf remote-terminal"))
+        XCTAssertTrue(arguments.contains("__QUILLCODE_TERMINAL_"))
+    }
+
+    func testTerminalCommandPersistsSSHRemoteCWDAndEnvironment() async throws {
+        let root = try makeTempDirectory()
+        let remoteRoot = root.appendingPathComponent("remote repo")
+        try FileManager.default.createDirectory(at: remoteRoot, withIntermediateDirectories: true)
+        let argumentsFile = root.appendingPathComponent("ssh-args.txt")
+        let fakeSSH = try makeExecutingFakeSSH(in: root, argumentsFile: argumentsFile)
+        let connection = ProjectConnection.ssh(
+            path: remoteRoot.path,
+            host: "feather.local",
+            user: "quill"
+        )
+        let project = ProjectRef(name: "Feather", path: connection.path, connection: connection)
+        let model = QuillCodeWorkspaceModel(
+            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
+            sshRemoteShellExecutor: SSHRemoteShellExecutor(
+                sshExecutable: fakeSSH.path,
+                connectTimeoutSeconds: 4
+            )
+        )
+
+        await model.runTerminalCommand(
+            "mkdir -p nested && cd nested && export QUILL_REMOTE_TERMINAL=works && printf remote-one",
+            workspaceRoot: root
+        )
+
+        let nestedPath = remoteRoot.appendingPathComponent("nested").path
+        XCTAssertEqual(model.terminal.entries[0].status, .done)
+        XCTAssertEqual(model.terminal.entries[0].stdout, "remote-one")
+        XCTAssertEqual(model.terminal.currentDirectoryPath, "ssh://quill@feather.local\(nestedPath)")
+        XCTAssertEqual(model.terminal.environmentOverrides["QUILL_REMOTE_TERMINAL"], "works")
+
+        await model.runTerminalCommand(
+            #"pwd && printf ':' && printf "$QUILL_REMOTE_TERMINAL""#,
+            workspaceRoot: root
+        )
+
+        XCTAssertEqual(model.terminal.entries[1].status, .done)
+        XCTAssertEqual(model.terminal.entries[1].stdout, "\(nestedPath)\n:works")
+        let arguments = try String(contentsOf: argumentsFile, encoding: .utf8)
+        XCTAssertTrue(arguments.contains("cd '\(nestedPath.replacingOccurrences(of: "'", with: "'\\''"))' &&"))
     }
 
     func testTerminalCommandAppearsAsRunningBeforeCompletion() async throws {
