+++ docs/DECISIONS.md
@@ -219,7 +219,7 @@
 - Workspace slash-command integration coverage lives in `WorkspaceSlashCommandIntegrationTests`, not `WorkspaceModelTests`. Command-palette slash prefills, core slash dispatch, local environment action slash execution, local model/mode/thread lifecycle commands, context compaction, and status transcript assertions cross composer submission, command dispatch, workspace surfaces, and transcript creation; keep those flows together while pure transcript copy stays in `WorkspaceSlashCommandTranscriptPlannerTests` and remote/automation-specific slash flows stay with their broader owning features until extracted.
 - Workspace local environment integration coverage lives in `WorkspaceLocalEnvironmentIntegrationTests`, not `WorkspaceModelTests`. Local environment action loading into project state, command-palette execution, metadata-backed environment redaction, bounded working directories, and timeouts cross the workspace model, metadata loader, shell-call planner, and tool cards; keep those flows together outside the model monolith while direct loader behavior stays focused. `/env` transcript behavior belongs to `WorkspaceSlashCommandIntegrationTests` with the rest of slash-command routing.
 - Workspace automation integration coverage lives in `WorkspaceAutomationIntegrationTests`, not `WorkspaceModelTests`. Automation persistence, create/pause/resume/delete commands, scheduled and natural-language follow-ups/workspace checks, slash scheduling commands, due automation execution, recurrence advancement, and run reports cross workspace model, stores, surfaces, transcript creation, and thread persistence; keep those flows together while pure automation record and run-draft logic stays in `WorkspaceAutomationEngineTests`.
-- Workspace terminal integration coverage lives in `WorkspaceTerminalIntegrationTests`, not `WorkspaceModelTests`. Local terminal execution, streaming output, cwd/environment persistence, clear-history behavior, selected-project resets, cancellation, and stop-all behavior cross the workspace model, terminal engine, shell runner, terminal surface, and async task lifecycle; keep those flows together outside the model monolith while pure terminal state transitions stay in `WorkspaceTerminalEngineTests`. SSH terminal integration still lives with remote-project tests until fake SSH support is shared.
+- Workspace terminal integration coverage lives in `WorkspaceTerminalIntegrationTests`, not `WorkspaceModelTests`. Local terminal execution, SSH Remote terminal execution, streaming output, cwd/environment persistence, clear-history behavior, selected-project resets, cancellation, and stop-all behavior cross the workspace model, terminal engine, shell/SSH runners, terminal surface, and async task lifecycle; keep those flows together outside the model monolith while pure terminal state transitions stay in `WorkspaceTerminalEngineTests`.
 - Shared workspace integration fixtures live in `WorkspaceModelIntegrationTestSupport` only while they are genuinely cross-domain. Fake SSH, fake GitHub CLI, temporary git repositories, and fixed/recording LLM clients are common enough to share, but new broad helpers should move to domain-specific support files once they become owned by one integration cluster.
 - Worktree request values, tool-call construction, and handoff transcript records live outside `WorkspaceModel`. `WorkspaceWorktreeRequests` owns the public create/remove request structs used by dialogs, SwiftUI, and desktop controllers. `WorkspaceWorktreeToolCallPlanner` owns canonical `host.git.worktree.create` and `host.git.worktree.remove` argument JSON, including branch/base trimming and force flags. `WorkspaceWorktreeOpenEngine` owns the pure local/SSH Remote `Worktree: ...` thread construction, notice payloads, and display labels. The workspace model owns only tool dispatch, local/SSH project registration, selection, persistence, and top-bar refresh after a worktree is created.
 - Workspace status copy and context labels live in `WorkspaceStatusTextBuilder`. `/status`, slash mode confirmations, top-bar subtitles, instruction labels, memory labels, and mode labels must use the same helper so transcript copy, SwiftUI surfaces, static HTML, and menu-bar surfaces do not drift. `WorkspaceModel` owns when to emit a status response; the builder owns what the status response says.
